\name{getDay}
\alias{getDay}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Get the day in a month (in a year) from days elapsed since 01-01-1970
}
\description{
 Function computes from days elapsed since 01-01-1970 the day in a month (in a year).
}
\usage{
getDay(daysSince01011970)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{daysSince01011970}{
    Days elapsed since 1970-01-01
}
}
\value{
 Day in a month (in a year) computed from days elapsed since 01-01-1970
}
\author{
Sabine Zinn
}
\examples{
getDay(2561)
}

