\name{getYear}
\alias{getYear}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Get the calendar year from days elapsed since 01-01-1970
}
\description{
 Function computes from days elapsed since 01-01-1970 the related calendar year.
}
\usage{
getYear(daysSince01011970)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{daysSince01011970}{
    Days elapsed since 1970-01-01
}
}
\value{
 Calendar year from days elapsed since 01-01-1970
}
\author{
Sabine Zinn
}
\examples{
getYear(2561)
}

