% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mort_es}
\alias{mort_es}
\title{Mortality Rates by Year, Age and Sex (Spain)}
\format{
A data frame with 12,876 rows and 5 variables:
\describe{
\item{year}{Calendar year.}
\item{age}{Age (character-coded).}
\item{female}{Female mortality rate (character-coded numeric).}
\item{male}{Male mortality rate (character-coded numeric).}
\item{total}{Total mortality rate (character-coded numeric).}
}
}
\usage{
mort_es
}
\description{
Mortality rates extracted from the HMD mortality database for Spain.
Contains period 1x1 mortality rates by age and sex.
}
\keyword{datasets}
