% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFctMicSim.r
\name{rate_cS}
\alias{rate_cS}
\title{rate_cS}
\usage{
rate_cS(allTr)
}
\arguments{
\item{allTr}{a matrix containing names of all transition rate functions.}
}
\value{
A dependency matrix
}
\description{
Helper function for MicSim.
A function for building a matrix containing the arguments of the transition rate functions.
}
\details{
Construct matrix noting dependencies of functions
Build dependency matrix
}
\keyword{internal}
