% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rates_mdd}
\alias{rates_mdd}
\title{Rate Data for Simulation of MDD example}
\format{
\subsection{\code{rates_mdd}}{

A data frame with 49 rows and 17 columns:
\describe{
\item{age}{Age of individual}
\item{incidence_lo_m}{Incidence Rates low education men}
\item{cf_incidence_lo_f}{Counterfactual Incidence Rate low female}
...
}
}
}
\source{
https://doi.org/10.1093/eurpub/ckae066
}
\usage{
rates_mdd
}
\description{
Data used to construct transition rates from @lepe2024
}
\keyword{datasets}
